Provide a way to get a version and hash of the build.
This commit provides a simple subcommand `version` to print the version and hash of the repo when the binary was built.

Note that this does not yet commit to true versioning of the tool. The version is left as "x.x.x" for now until we really commit to semantic versioning.

Having the hash of the build will help for debugging to know what version of the tool is being used. See for example issue #5.